+++ CHANGELOG.rst
@@ -1,3 +1,10 @@
+UNRELEASED
+----------
+
+- Fixed exception handling so they are properly cleared in Python 3.12, due to the new `sys.last_exc <https://docs.python.org/3/library/sys.html#sys.last_exc>`__ attribute (`#532`_).
+
+..  _#532: https://github.com/pytest-dev/pytest-qt/issues/532
+
 4.3.1 (2023-12-22)
 ------------------